Feat add agent tools (#47)
PR expanded to include tools for extracting ticket and invoice metadata. Workflow also includes a final tool node to structure the final output in a way that we can either consume downstream or expand to update a table in Snowflake directly.

Python connection is now once per ticket being processed as opposed to once per tool call. It's passed at runtime as a context object to the graph.