[InvoiceIQ] Create Invoice Table and Sync with File Metadata (#56)
As discussed with Jason and Seth, this creates
`INVOICEIQ.SERVICE.INVOICES` table where each row tracks an invoice and
is linked to the row of the corresponding file in the file metadata
table.

This way, the file metadata table will be modified by AI extraction
tasks, and the status there will track if the extraction has succeeded
or is still pending.

The INVOICES table will contain the same file metadata information, but
will be used to track the actual approval or rejection of the invoice.
This means that the agent you're working on will need to operate on this
INVOICES table and update the approval/rejection/pending status there.
Also, through the UI, users will be able to update this table and then
retrigger the agent workflow for that invoice (since updating some field
is not guaranteed to cause the invoice to be approved).