Add AI workflow tool to vibe-code workflows (#63)
Initial check-in of the AI workflows tool. The goal is for this to be a
vibe-code development tool for workflows. The user will be able to
express what they want in natural language. There will be a no-code
graphical view (with perhaps no-code editing tools later), but they'll
be able to see the code.

This push is a functioning frontend with a fake backend. The frontend
resembles VSCode with a file browser on the left, a code pane, a
workflow viewer pane, and an AI chat interface. The code editor uses
microsoft-monaco which is from the same source as VSCode, meaning we get
syntax highlighting and can easily add other features.

Next step is to implement the backend by hooking it up to Snova.