Jsummer/align agent invoice table (#62)
PR does the following:

- Switches from processing tickets to invoices. The process is as
follows:
1. All invoices captured in STREAM are placed in STAGING table via a
Task. This is necessary as once a record is consumed. the STREAM
refreshes its pointer so we are unable to process a single invoice from
the STREAM without wiping out the entire queue.
2. Using a separate task, individual records from the STAGING table are
passed to the agent.
3. Within the task, the agent result is used to update INVOICES table.
Note that the agent doesn't do this directly due to the issue cited in
#1 and the fact that this creates an infinite loop.
4. The invoice is removed from the STAGING table.

- Revised the used of ChatOpenAI url and credentials by using the
container token directly. External access is also no longer needed.