[sales-ai-platform] potential new use cases table inserts (#232)
We created a table with a separate row for each new extracted use case.
This table contains a status column that reflects if a record has been
rejected or accepted.

We add checks in the workflow node that inserts new use cases into this
table. The checks first look to see if the same call has already been
analyzed with records already in the new use cases table. Secondly, we
employ an LLM to judge if the new use case is essentially a duplicate of
an existing use case. If the call has already been analyzed or if there
is a probable duplicate, the new use case is not inserted into the
table.