[sales-ai-platform] comment out comparison against current sfdc use cases (#255)
We no longer seem to have access to the SALES.KNOWLEDGE_ASSISTANT schema
which contains VH_DELIVERABLE_C. This table contains current SFDC use
cases. So for the time being we comment out the code that compares a
proposed use case to the current use cases in that table.

We are still comparing to the use cases that exist in the Proposed New
Use Cases table (POTENTIAL_NEW_USE_CASES).

This deduping process leverages another LLM, so the decision needs to be
made if it is a desired step or if it is preferred to re-show potential
use cases to the account team.