[knowledge-builder] feat: add initial snowpark notebook demo as script (#275)
I copied over the code we started working on Friday in the
TWHITE_SABBAS_2025-11-21 11:45:57 notebook.

I made a few improvements since then given some lessons learned, mostly:

- organizing code, created ingest function
- using
[`snowflake.snowpark.DataFrameAIFunctions.split_text_recursive_character`](https://docs.snowflake.com/en/developer-guide/snowpark/reference/python/latest/snowpark/api/snowflake.snowpark.DataFrameAIFunctions.split_text_recursive_character)

edit: this currently depends on that "kb_knowledge.csv" file. I would
also like to blend it with the existing scripts if we need to. :)