[sales-ai-platform] make it easy to turn off the LLM dedup process (#292)
add an env variable (DEDUP_NEW_USE_CASES) that can be set to true or
false. If true, then an LLM is used to see if the proposed new use case
exists in the new use case table already. If false, then the LLM is not
used and all new use case records are inserted into the table, even if
they already exist from a previous call.