p67: v0 of injecting the sdk object into the workflow. (#370)
New change requires that the workflow have a manifest.yaml which is then
loaded, and understood into a config that is used to hydrate a SDK
object. That
object is then injected into the workflow. The consequences is that the
user
has access to things like querying Snowflake, but controlled by our
system.

TODO: still need to hydrate from secrets instead of just having them
defined in the manifest.yaml